s2/requirements.txt
# requirements.txt para o serviço s2

fastapi
uvicorn
kafka-python
pymongo
redis
psycopg2-binary


